requirements.txt
streamlit>=1.28
langchain>=0.3.0
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-pinecone
openai>=1.2
google-generativeai
firebase-admin
duckduckgo-search
anthropic>=0.3.0
trubrics>=1.4.3
streamlit-feedback
pinecone-client
pypdf
python-dotenv
